requirements.txt
dataclasses==0.6
Django==2.2.5
django-elasticsearch-dsl==7.0.0
django-extensions==2.2.1
django-filter==2.2.0
django-nose==1.4.6
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djongo==1.2.35
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
Faker==2.0.1
mongoengine==0.18.2
nose==1.3.7
Pillow==6.1.0
PyJWT==1.7.1
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.25.3
